requirements.txt
Django==5.2.6
djangorestframework==3.14.0
oracledb==2.4.1
Pillow==10.4.0
python-decouple==3.8
